Async2

+++ MovieStoreTISAI/MovieStoreTISAI.BL/Services/MoviesService.cs
@@ -22,36 +22,33 @@ namespace MovieStoreTISAI.BL.Services
         }
 
 
-        public void Add(Movie movie)
+        public async Task Add(Movie movie)
         {
-            if (movie == null|| movie.Actors == null) return;
-            foreach (var actor in movie.Actors)
-            {
-                //if (Guid.TryParse(actor.Id, out _)) return;
-            }
-            //  _movieRepository.Add(movie);
+           
+             await _movieRepository.Add(movie);
         }
-        public void Delete(string id)
+        public async Task Delete(string id)
         {
-            _movieRepository.Delete(id);
+            if (string.IsNullOrWhiteSpace(id)) return;
+           await _movieRepository.Delete(id);
         }
 
-        public Movie? GetByID(string id)
+        public async Task<Movie?> GetByID(string id)
         {
             if (string.IsNullOrEmpty(id))
             {
                 _logger.LogError("MovieId is null or empty");
                 return null;
             }
-            return _movieRepository.GetByID(id);
+            return  await _movieRepository.GetByID(id);
         }
 
-        public List<Movie> GetAll()
+        public async Task<List<Movie>> GetAll()
         {
-            return _movieRepository.GetAll();
+            return await _movieRepository.GetAll();
         }
 
-        public void AddActorToMovie(string movieId, string actorId)
+        public async Task AddActorToMovie(string movieId, string actorId)
         {
             if (string.IsNullOrEmpty(movieId) || string.IsNullOrEmpty(actorId))
             {
@@ -59,13 +56,13 @@ namespace MovieStoreTISAI.BL.Services
                 return;
             }
 
-            if(!Guid.TryParse(movieId, out _) || !Guid.TryParse(actorId, out _))
+            if (!Guid.TryParse(movieId, out _) || !Guid.TryParse(actorId, out _))
             {
                 _logger.LogError("MovieId or ActorId is not a valid Guid");
                 return;
             }
 
-            var movie = _movieRepository.GetByID(movieId);
+            var movie = await _movieRepository.GetByID(movieId);
 
             if (movie == null)
             {
@@ -73,9 +70,9 @@ namespace MovieStoreTISAI.BL.Services
                 return;
             }
 
-            var actor = _actorRepository.GetByID(actorId);
-            if (actor == null) 
-            { 
+            var actor = await _actorRepository.GetByID(actorId);
+            if (actor == null)
+            {
                 _logger.LogError("Actor not found");
                 return;
             }
@@ -83,9 +80,14 @@ namespace MovieStoreTISAI.BL.Services
             {
                 movie.Actors = new List<string>();
             }
-            movie.Actors.Add(actorId); 
+            movie.Actors.Add(actorId);
 
-            _movieRepository.Update(movie);
+           await _movieRepository.Update(movie);
+        }
+
+        public async Task Update(Movie movie)
+        {
+            await _movieRepository.Update(movie);
         }
     }
 }
